requirements.txt
python-dotenv
requests
python-socketio
events 